test: split unit tests into separate files to make coverage rate more accurate
