+++ src/protocols/light_client/components/mod.rs
@@ -2,6 +2,9 @@ mod send_block_proof;
 mod send_block_samples;
 mod send_last_state;
 
+#[cfg(test)]
+mod tests;
+
 pub(crate) use send_block_proof::SendBlockProofProcess;
 pub(crate) use send_block_samples::SendBlockSamplesProcess;
 pub(crate) use send_last_state::SendLastStateProcess;
