+++ src/protocols/light_client/components/tests/mod.rs
@@ -1,0 +1,1 @@
+mod send_block_samples;
