+++ src/protocols/light_client/tests/mod.rs
@@ -1,0 +1,1 @@
+mod sampling;
