+++ src/tests/mod.rs
@@ -1,0 +1,2 @@
+mod service;
+mod verify;
